pyproject.toml
[project]
name = "workspace"
version = "0.1.0"
description = ""
authors = [
    {name = "Yuji-github",email = "81yishikawa@gmail.com"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "dash (>=3.3.0,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "yfinance (>=1.0,<2.0)",
    "pyarrow (>=22.0.0,<23.0.0)",
    "google-genai (>=1.56.0,<2.0.0)",
    "boto3 (>=1.42.21,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "debugpy (>=1.8.19,<2.0.0)",
    "ipykernel (>=7.1.0,<8.0.0)",
    "black (>=25.12.0,<26.0.0)",
]
